change the way of specifying properties of entities. (#93)
* change the way of specifying properties of entities.

Properties are now added directly to the class instead of to the instance in
__init__. The mechanism is via metaclass, similar to that used by several python
ORMS.

A new property creating class is provided for this purpose.

Properties can now specify "backrefs" in a similar manner to backrefs in
SQLAlchemy.

The important acheivement is that backrefs allow for specifying bidirectional
relationships without increasingly complicated juggling of recursive imports.
The modules and imports may be arranged as a simple spanning tree. The side of
a relationship that does not import it's remote end, may have it's relationship
specified on the remote via a backref.

This is useful in the exptected upcoming case of domain tags, where it is
important to be able to access tags from any direction.

* Rename _is_entity tag to more obscure _is_merchi_entity to make name
collisions less likely.