proposed method for teaching typescript SDK about the attributes owned by (#181)
each type of entity.

I tried to strike a balance between using advanced/new/experimental features/
writing complex code, vs the crazy boilerplate if everything is entirely
manual. That lead me to use decorators to collect the list of
attributes, but to give up on the more adventurous plan of using
typescript transfomers to produce the properties themselves.